+++ apps/ledger-live-desktop/src/newArch/features/DynamicContent/__integrations__/PortfolioContentCards.test.tsx
@@ -107,7 +107,10 @@ describe("PortfolioContentCards", () => {
     expect(track).not.toHaveBeenCalledWith("contentcard_clicked", expect.any(Object));
     act(() => cta0.click());
     expect(logContentCardClick).toHaveBeenCalledTimes(1);
-    expect(logContentCardClick).toHaveBeenCalledWith(asBrazeCard(Cards[0]));
+    expect(logContentCardClick).toHaveBeenCalledWith({
+      ...asBrazeCard(Cards[0]),
+      url: Cards[0].id,
+    });
     expect(track).toHaveBeenCalledWith("contentcard_clicked", {
       contentcard: "Foo",
       link: "ledger-live://deep-link",