+++ libs/ledger-live-common/src/e2e/enum/Fee.ts
@@ -1,5 +1,5 @@
 export enum Fee {
-  FAST = "fast",
-  MEDIUM = "medium",
-  SLOW = "slow",
+  FAST = "Fast",
+  MEDIUM = "Medium",
+  SLOW = "Slow",
 }