+++ libs/live-nft-react/CHANGELOG.md
@@ -1,5 +1,21 @@
 # @ledgerhq/live-nft-react
 
+## 0.5.6
+
+### Patch Changes
+
+- Updated dependencies [[`5e18866`](https://github.com/LedgerHQ/ledger-live/commit/5e18866320b843632699659ee66f6c410c108c1e)]:
+  - @ledgerhq/coin-framework@1.0.0
+  - @ledgerhq/live-nft@0.5.6
+
+## 0.5.6-next.0
+
+### Patch Changes
+
+- Updated dependencies [[`5e18866`](https://github.com/LedgerHQ/ledger-live/commit/5e18866320b843632699659ee66f6c410c108c1e)]:
+  - @ledgerhq/coin-framework@1.0.0-next.0
+  - @ledgerhq/live-nft@0.5.6-next.0
+
 ## 0.5.5
 
 ### Patch Changes